requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==1.7.1
python-dotenv==0.21.0
python-environ==0.4.54
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.38.3
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
